backend/requirements.txt
# FastAPI + ASGI
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Async HTTP client (fetching from IETF Datatracker + rfc-editor.org)
httpx==0.28.1

# Environment variable loading
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx  # also used as async test client for FastAPI

# Optional: for future concurrency / background tasks
anyio==4.7.0

# Edge TTS — free neural text-to-speech via Microsoft Edge
edge-tts>=7.0.0
